Implemented most of the memory instructions

+++ sgemew/src/hardware/MemoryDevice.cpp
@@ -23,7 +23,7 @@ MemoryDevice::MemoryDevice()
 {
 }
 
-MemoryDevice::MemoryDevice(uint32_t size)
+MemoryDevice::MemoryDevice(uint64_t size)
 {
 	initMemory(size);
 }
@@ -33,19 +33,19 @@ MemoryDevice::~MemoryDevice()
 	if (initialized) { delete[] mem; }
 }
 
-uint8_t MemoryDevice::read8(uint32_t addr)
+uint8_t MemoryDevice::read8(uint64_t addr)
 {
 	return (*this)[addr];
 }
 
-uint16_t MemoryDevice::read16(uint32_t addr)
+uint16_t MemoryDevice::read16(uint64_t addr)
 {
 	uint16_t r = (*this)[addr] << 8;
 	r |= (*this)[addr + 1];
 	return r;
 }
 
-uint32_t MemoryDevice::read32(uint32_t addr)
+uint32_t MemoryDevice::read32(uint64_t addr)
 {
 	uint32_t r = (*this)[addr] << 24;
 	r |= (*this)[addr + 1] << 16;
@@ -54,7 +54,7 @@ uint32_t MemoryDevice::read32(uint32_t addr)
 	return r;
 }
 
-uint64_t MemoryDevice::read64(uint32_t addr)
+uint64_t MemoryDevice::read64(uint64_t addr)
 {
 	uint64_t r = (uint64_t)(*this)[addr] << 56;
 	r |= (uint64_t)(*this)[addr + 1] << 48;
@@ -67,18 +67,18 @@ uint64_t MemoryDevice::read64(uint32_t addr)
 	return r;
 }
 
-void MemoryDevice::write(uint32_t addr, uint8_t data)
+void MemoryDevice::write(uint64_t addr, uint8_t data)
 {
 	(*this)[addr] = data;
 }
 
-void MemoryDevice::write(uint32_t addr, uint16_t data)
+void MemoryDevice::write(uint64_t addr, uint16_t data)
 {
 	(*this)[addr] = 0xff & (data >> 8);
 	(*this)[addr + 1] = 0xff & data;
 }
 
-void MemoryDevice::write(uint32_t addr, uint32_t data)
+void MemoryDevice::write(uint64_t addr, uint32_t data)
 {
 	(*this)[addr] = 0xff & (data >> 24);
 	(*this)[addr + 1] = 0xff & (data >> 16);
@@ -86,7 +86,7 @@ void MemoryDevice::write(uint32_t addr, uint32_t data)
 	(*this)[addr + 3] = 0xff & data;
 }
 
-void MemoryDevice::write(uint32_t addr, uint64_t data)
+void MemoryDevice::write(uint64_t addr, uint64_t data)
 {
 	(*this)[addr] = 0xff & (data >> 56);
 	(*this)[addr + 1] = 0xff & (data >> 48);
@@ -98,12 +98,12 @@ void MemoryDevice::write(uint32_t addr, uint64_t data)
 	(*this)[addr + 7] = 0xff & data;
 }
 
-uint8_t& MemoryDevice::operator [](uint32_t addr)
+uint8_t& MemoryDevice::operator [](uint64_t addr)
 {
 	return mem[addr];
 }
 
-void MemoryDevice::initMemory(uint32_t size)
+void MemoryDevice::initMemory(uint64_t size)
 {
 	if (initialized) { return; }
 	aSize = size;
