+++ sgemew/src/hardware/MemoryDevice.h
@@ -40,69 +40,69 @@ public:
 	 * Creates a device with the specified amount of memory.
 	 * @param size Size of memory block to create.
 	 */
-	MemoryDevice(uint32_t size);
+	MemoryDevice(uint64_t size);
 	~MemoryDevice();
 	/**
 	 * Reads an 8-bit number
 	 * @param addr Address to be read
 	 * @returns Value stored at specified address
 	 */
-	virtual uint8_t read8(uint32_t addr);
+	virtual uint8_t read8(uint64_t addr);
 	/**
 	 * Reads a 16-bit big-endian number
 	 * @param addr Address to start reading
 	 * @returns Value stored at specified address
 	 */
-	virtual uint16_t read16(uint32_t addr);
+	virtual uint16_t read16(uint64_t addr);
 	/**
 	 * Reads a 32-bit big-endian number
 	 * @param addr Address to start reading
 	 * @returns Value stored at specified address
 	 */
-	virtual uint32_t read32(uint32_t addr);
+	virtual uint32_t read32(uint64_t addr);
 	/**
 	 * Reads a 64-bit big-endian number
 	 * @param addr Address to start reading
 	 * @returns Value stored at specified address
 	 */
-	virtual uint64_t read64(uint32_t addr);
+	virtual uint64_t read64(uint64_t addr);
 
 	/**
 	 * Writes a big-endian number
 	 * @param addr Address to write
 	 * @param data Number to write
 	 */
-	virtual void write(uint32_t addr, uint8_t data);
+	virtual void write(uint64_t addr, uint8_t data);
 	/**
 	 * Writes a big-endian number
 	 * @param addr Address to write
 	 * @param data Number to write
 	 */
-	virtual void write(uint32_t addr, uint16_t data);
+	virtual void write(uint64_t addr, uint16_t data);
 	/**
 	 * Writes a big-endian number
 	 * @param addr Address to write
 	 * @param data Number to write
 	 */
-	virtual void write(uint32_t addr, uint32_t data);
+	virtual void write(uint64_t addr, uint32_t data);
 	/**
 	 * Writes a big-endian number
 	 * @param addr Address to write
 	 * @param data Number to write
 	 */
-	virtual void write(uint32_t addr, uint64_t data);
+	virtual void write(uint64_t addr, uint64_t data);
 
 	/**
 	 * Allows array-style access to memory block.
 	 * @param addr Address to access
 	 */
-	virtual uint8_t &operator[](uint32_t addr);
+	virtual uint8_t& operator[](uint64_t addr);
 
 private:
 	bool initialized = false;
 	uint32_t aSize;
 	uint8_t *mem;
-	virtual void initMemory(uint32_t size);
+	virtual void initMemory(uint64_t size);
 
 };
